+++ routes/api.php
@@ -5,6 +5,7 @@ use App\Http\Controllers\Api\Admin\ProductController as ApiAdminProductControlle
 use Illuminate\Support\Facades\Route;
 
 Route::middleware('admin.api')->prefix('admin')->name('api.admin.')->group(function () {
+    Route::get('/products', [ApiAdminProductController::class, 'index'])->name('products.index');
     Route::get('/products/{product}', [ApiAdminProductController::class, 'show'])->name('products.show');
     Route::patch('/products/{product}', [ApiAdminProductController::class, 'update'])->name('products.update');
     Route::post('/orders', [ApiAdminOrderController::class, 'createDraft'])->name('orders.store');